+++ src/plugins/data_source_management/public/components/data_source_selectable/data_source_selectable.test.tsx
@@ -12,8 +12,6 @@ import { AuthType } from '../../types';
 import { getDataSourcesWithFieldsResponse, mockResponseForSavedObjectsCalls } from '../../mocks';
 import { render } from '@testing-library/react';
 import * as utils from '../utils';
-import { EuiSelectable } from '@elastic/eui';
-import { dataSourceOptionGroupLabel } from '../utils';
 
 describe('DataSourceSelectable', () => {
   let component: ShallowWrapper<any, Readonly<{}>, React.Component<{}, {}, any>>;
@@ -103,7 +101,7 @@ describe('DataSourceSelectable', () => {
 
     await nextTick();
 
-    const button = await container.findByTestId('dataSourceSelectableContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceSelectableButton');
     button.click();
 
     expect(container.getByTestId('dataSourceSelectableContextMenuPopover')).toBeVisible();
@@ -192,7 +190,7 @@ describe('DataSourceSelectable', () => {
     );
 
     await nextTick();
-    const button = await container.findByTestId('dataSourceSelectableContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceSelectableButton');
     expect(button).toHaveTextContent('test2');
   });
 
@@ -211,7 +209,7 @@ describe('DataSourceSelectable', () => {
       />
     );
     await nextTick();
-    const button = await container.findByTestId('dataSourceSelectableContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceSelectableButton');
     expect(button).toHaveTextContent('test2');
   });
 
@@ -230,7 +228,7 @@ describe('DataSourceSelectable', () => {
       />
     );
     await nextTick();
-    const button = await container.findByTestId('dataSourceSelectableContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceSelectableButton');
     expect(button).toHaveTextContent('');
     expect(toasts.addWarning).toBeCalledWith('Data source with id: undefined is not available');
   });
@@ -250,7 +248,7 @@ describe('DataSourceSelectable', () => {
       />
     );
     await nextTick();
-    const button = await container.findByTestId('dataSourceSelectableContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceSelectableButton');
     expect(button).toHaveTextContent('');
     expect(toasts.addWarning).toBeCalledWith('Data source with id: undefined is not available');
   });
@@ -403,49 +401,9 @@ describe('DataSourceSelectable', () => {
     expect(onSelectedDataSource).toBeCalledWith([{ id: 'test2', label: 'test2' }]);
     expect(onSelectedDataSource).toHaveBeenCalled();
   });
-
-  it('should render opensearch cluster group label at the top of options, when there are options availiable', async () => {
+  it('should render no data source when no data source filtered out and hide local cluster', async () => {
     const onSelectedDataSource = jest.fn();
-    component = shallow(
-      <DataSourceSelectable
-        savedObjectsClient={client}
-        notifications={toasts}
-        onSelectedDataSources={onSelectedDataSource}
-        disabled={false}
-        hideLocalCluster={false}
-        fullWidth={false}
-      />
-    );
-
-    component.instance().componentDidMount!();
-    await nextTick();
-    const optionsProp = component.find(EuiSelectable).prop('options');
-    expect(optionsProp[0]).toEqual(dataSourceOptionGroupLabel.opensearchCluster);
-  });
-
-  it('should not render opensearch cluster group label, when there is no option availiable', async () => {
-    const onSelectedDataSource = jest.fn();
-    spyOn(utils, 'getDefaultDataSource').and.returnValue([]);
-    component = shallow(
-      <DataSourceSelectable
-        savedObjectsClient={client}
-        notifications={toasts}
-        onSelectedDataSources={onSelectedDataSource}
-        disabled={false}
-        hideLocalCluster={false}
-        fullWidth={false}
-      />
-    );
-
-    component.instance().componentDidMount!();
-    await nextTick();
-    const optionsProp = component.find(EuiSelectable).prop('options');
-    expect(optionsProp).toEqual([]);
-  });
-
-  it('should render group lablel normally after onChange', async () => {
-    const onSelectedDataSource = jest.fn();
-    component = shallow(
+    render(
       <DataSourceSelectable
         savedObjectsClient={client}
         notifications={toasts}
@@ -453,39 +411,12 @@ describe('DataSourceSelectable', () => {
         disabled={false}
         hideLocalCluster={true}
         fullWidth={false}
-        selectedOption={[{ id: 'test1', label: 'test1' }]}
+        selectedOption={[{ id: 'test2' }]}
+        dataSourceFilter={(ds) => false}
       />
     );
-    const componentInstance = component.instance();
-
-    componentInstance.componentDidMount!();
     await nextTick();
-    const optionsPropBefore = component.find(EuiSelectable).prop('options');
-    expect(optionsPropBefore).toEqual([
-      dataSourceOptionGroupLabel.opensearchCluster,
-      {
-        id: 'test1',
-        label: 'test1',
-        checked: 'on',
-      },
-      {
-        id: 'test2',
-        label: 'test2',
-      },
-      {
-        id: 'test3',
-        label: 'test3',
-      },
-    ]);
-    componentInstance.onChange([
-      dataSourceOptionGroupLabel.opensearchCluster,
-      { id: 'test2', label: 'test2', checked: 'on' },
-    ]);
-    await nextTick();
-    const optionsPropAfter = component.find(EuiSelectable).prop('options');
-    expect(optionsPropAfter).toEqual([
-      dataSourceOptionGroupLabel.opensearchCluster,
-      { id: 'test2', label: 'test2', checked: 'on' },
-    ]);
+    expect(toasts.add).toBeCalled();
+    expect(onSelectedDataSource).toBeCalledWith([]);
   });
 });