+++ src/plugins/vis_type_timeseries/server/plugin.ts
@@ -51,6 +51,12 @@ import { visDataRoutes } from './routes/vis';
 import { fieldsRoutes } from './routes/fields';
 import { SearchStrategyRegistry } from './lib/search_strategies';
 import { uiSettings } from './ui_settings';
+import { setDataSourceEnabled } from './lib/services';
+import {
+  TIMESERIES_VISUALIZATION_CLIENT_WRAPPER_ID,
+  timeSeriesVisualizationClientWrapper,
+  TIMESERIES_VISUALIZATION_CLIENT_WRAPPER_PRIORITY,
+} from './lib/timeseries_visualization_client_wrapper';
 
 export interface LegacySetup {
   server: Server;
@@ -114,6 +120,13 @@ export class VisTypeTimeseriesPlugin implements Plugin<VisTypeTimeseriesSetup> {
       router,
       searchStrategyRegistry,
     };
+
+    setDataSourceEnabled({ enabled: !!plugins.dataSource });
+    core.savedObjects.addClientWrapper(
+      TIMESERIES_VISUALIZATION_CLIENT_WRAPPER_PRIORITY,
+      TIMESERIES_VISUALIZATION_CLIENT_WRAPPER_ID,
+      timeSeriesVisualizationClientWrapper
+    );
 
     (async () => {
       const validationTelemetry = await this.validationTelementryService.setup(core, {